# openapi.yaml
openapi: 3.1.0
info:
  title: RescueBite API
  description: |
    Food Waste Reduction Marketplace REST API

    Built with FastAPI + SQLModel + PostgreSQL + Redis + Celery.

    **Auth flow:** Register → Verify Email (OTP) → Login → Bearer token → Refresh → Logout

    **Business flow:**
    1. Vendor registers and submits business profile
    2. Admin approves vendor
    3. Vendor creates listings (food items with discount)
    4. Customer browses listings, places order
    5. Customer pays via Kaspi Pay (or simulated)
    6. Vendor confirms and marks order ready for pickup
    7. Customer picks up with token
    8. Celery Beat applies price decay every 72h
  version: 1.0.0
  contact:
    name: RescueBite Team
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /listings:
    get:
      tags: [Listings]
      summary: Browse active food listings
      operationId: listListings
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: Cursor for pagination
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: lat
          in: query
          schema: { type: number, format: float }
          description: Latitude for geo-filtering
        - name: lng
          in: query
          schema: { type: number, format: float }
          description: Longitude for geo-filtering
      responses:
        "200":
          description: Paginated listing list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPage'

    post:
      tags: [Listings]
      summary: Create a new food listing (vendor only)
      operationId: createListing
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListingCreate' }
            example:
              title: "Свежие суши (конец дня)"
              description: "Не раскупленные суши от закрытия — 50% скидка"
              original_price: 3500
              discount_percentage: 50
              quantity_total: 10
              pickup_window_start: "2025-05-15T18:00:00"
              pickup_window_end: "2025-05-15T21:00:00"
              allergens: [fish, gluten]
              latitude: 51.180
              longitude: 71.446
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /listings/allergen-check:
    post:
      tags: [Listings]
      summary: Check ingredients against user allergen profile
      operationId: allergenCheck
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ingredients, user_allergens]
              properties:
                ingredients:
                  type: array
                  items: { type: string }
                  example: ["мука", "молоко", "яйца"]
                user_allergens:
                  type: array
                  items: { $ref: '#/components/schemas/AllergenCode' }
      responses:
        "200":
          description: Allergen analysis result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllergenFilterResponse' }
  /listings/vendor/my-listings:
    get:
      tags: [Listings]
      summary: Get authenticated vendor's own listings
      operationId: myListings
      security: [{ BearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ListingStatus' }
      responses:
        "200":
          description: Vendor's listing page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingPage' }
  /listings/{listing_id}:
    get:
      tags: [Listings]
      summary: Get a single listing by ID
      operationId: getListing
      parameters:
        - name: listing_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingResponse' }
        "404": { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Listings]
      summary: Update own listing (vendor or admin)
      operationId: updateListing
      security: [{ BearerAuth: [] }]
      parameters:
        - name: listing_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListingUpdate' }
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingResponse' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Listings]
      summary: Delete own listing
      operationId: deleteListing
      security: [{ BearerAuth: [] }]
      parameters:
        - name: listing_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Listing deleted
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
...
  /admin/listings:
    get:
      tags: [Admin]
      summary: List all listings (all statuses)
      operationId: adminListListings
      security: [{ BearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ListingStatus' }
        - name: vendor_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Listing page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingPage' }
  /admin/listings/{listing_id}:
    patch:
      tags: [Admin]
      summary: Update any listing
      operationId: adminUpdateListing
      security: [{ BearerAuth: [] }]
      parameters:
        - name: listing_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ListingUpdate' }
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingResponse' }

    delete:
      tags: [Admin]
      summary: Delete listing
      operationId: adminDeleteListing
      security: [{ BearerAuth: [] }]
      parameters:
        - name: listing_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Listing deleted
        "409": { $ref: '#/components/responses/Conflict' }
...
  /jobs/status:
    get:
      tags: [Background Jobs]
      summary: Celery queue overview (admin only)
      operationId: jobsOverview
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Queue stats and active tasks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobsOverview' }
        "503":
          description: Cannot reach Celery workers
  /jobs/{task_id}:
    get:
      tags: [Background Jobs]
      summary: Get specific task result by ID
      operationId: getJob
      security: [{ BearerAuth: [] }]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobInfo' }
        "404": { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
    Forbidden:
      description: Insufficient role or ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
    Conflict:
      description: Conflict — duplicate resource or business rule violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
...
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail: { type: string }
...
    ListingStatus:
      type: string
      enum: [draft, active, discounted, free, compost, paused, sold_out]
...
    AllergenCode:
      type: string
      enum: [none, gluten, dairy, eggs, nuts, soy, fish, shellfish, sesame]
...
    ListingCreate:
      type: object
      required: [title, description, original_price, discount_percentage, quantity_total, pickup_window_start, pickup_window_end, allergens, latitude, longitude]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 1000 }
        original_price: { type: integer, minimum: 500, description: Price in KZT }
        discount_percentage: { type: number, minimum: 1, maximum: 90 }
        quantity_total: { type: integer, minimum: 1 }
        pickup_window_start: { type: string, format: date-time }
        pickup_window_end: { type: string, format: date-time }
        allergens:
          type: array
          items: { $ref: '#/components/schemas/AllergenCode' }
          minItems: 1
        latitude: { type: number }
        longitude: { type: number }
        photo_url: { type: string, nullable: true }
    ListingUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        quantity_total: { type: integer, minimum: 1 }
        pickup_window_start: { type: string, format: date-time }
        pickup_window_end: { type: string, format: date-time }
        allergens:
          type: array
          items: { $ref: '#/components/schemas/AllergenCode' }
        status: { $ref: '#/components/schemas/ListingStatus' }
    ListingResponse:
      type: object
      properties:
        id: { type: integer }
        vendor_id: { type: integer }
        title: { type: string }
        description: { type: string }
        original_price: { type: integer }
        current_price: { type: integer }
        discount_percentage: { type: number }
        quantity_total: { type: integer }
        quantity_available: { type: integer }
        status: { $ref: '#/components/schemas/ListingStatus' }
        pickup_window_start: { type: string, format: date-time }
        pickup_window_end: { type: string, format: date-time }
        allergens:
          type: array
          items: { $ref: '#/components/schemas/AllergenCode' }
        latitude: { type: number }
        longitude: { type: number }
        photo_url: { type: string, nullable: true }
        days_active: { type: integer }
        created_at: { type: string, format: date-time }
    ListingPage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ListingResponse' }
        pagination: { $ref: '#/components/schemas/PaginationMeta' }
...
    PaginationMeta:
      type: object
      properties:
        limit: { type: integer }
        next_cursor: { type: string, nullable: true }
    AllergenFilterResponse:
      type: object
      properties:
        safe: { type: boolean }
        detected_allergens:
          type: array
          items: { $ref: '#/components/schemas/AllergenCode' }
        flagged_ingredients:
          type: array
          items: { type: string }
        message: { type: string }
    JobInfo:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        state: { type: string }
        result: { type: string, nullable: true }
        traceback: { type: string, nullable: true }
        date_done: { type: string, nullable: true }
    JobsOverview:
      type: object
      properties:
        queue_stats:
          type: object
          properties:
            active: { type: integer }
            reserved: { type: integer }
            scheduled: { type: integer }
            workers:
              type: array
              items: { type: string }
        active_tasks:
          type: array
          items: { type: object }
        scheduled_beats:
          type: array
          items: { type: object }